
\name{compareAVHRRimages}
\alias{compareAVHRRimages}

\title{
    Compares a single AVHRR image processed by different processing chains
}

\description{
This function reads the output from \link{readAVHRR} and compares AVHRR
images generated by different processing chains to visualize any important
differences in processing. An image is created.
}

\usage{
compareAVHRRimages(x, y, channel=1, map=TRUE, method="N")  
}

\arguments{
    \item{x}{R object (\link{list}) returned by \link{readAVHRR}}
    \item{channel}{channel identification}
    \item{map}{logical telling whether coast lines should be added}
    \item{merthod}{letter specifying how to combine channels ("D" is
    difference, "R" is ratio, "N" is none)}
}
\value{None}

\seealso{
    \link{readAVHRR}, \link{createAVHRRimage}, \link{compareAVHRRchains}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}
\note{
    The current version supports map overlay only for the standard map
    projection of the OSISAF HL centre. That is Polar Stereographic
    correct at 60 degrees North and not rotated.
}

\examples{
    # Read a AVHRR image from processing chain 1
    t1 <- readAVHRR(file="/data/avhrr/noaa18_03452_200610191134.h5")
    # Read a AVHRR image from processing chain 2
    t2 <- readAVHRR(file="/data/avhrr_old/noaa18_03452_200610191134_03456.h5")
    # Generate image on graphical device
    compareAVHRRimages(t1,t2)
}

\keyword{}
