\name{fm_findsollum}

\alias{fm_findsollum}

\title{
    Estimate AVHRR channel 3B TOA radiance
}

\description{
    Computes the solar radiance at top of atmosphere for CH3B, L3*cos(theta).
}

\usage{
    fm_findsollum(solang, satname, date)
}

\arguments{
    \item{solang}{Solar zenith angle}
    \item{satname}{Satellite name e.g. "NOAA-18"}
    \item{date}{Date identification}
}

\value{
    A value or a vector depending on the input
}

\seealso{
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
